+++ crates/karva_cli/tests/cli.rs
@@ -2533,96 +2533,6 @@ fn test_stdout_is_captured_and_displayed_with_args() -> anyhow::Result<()> {
     Ok(())
 }
 
-fn get_auto_use_kw(framework: &str) -> &str {
-    match framework {
-        "pytest" => "autouse",
-        "karva" => "auto_use",
-        _ => panic!("Invalid framework"),
-    }
-}
-
-#[rstest]
-fn test_auto_use_fixture(#[values("pytest", "karva")] framework: &str) -> anyhow::Result<()> {
-    let auto_use_kw = get_auto_use_kw(framework);
-
-    let test_code = format!(
-        r#"
-from {framework} import fixture
-
-@fixture
-def first_entry():
-    return "a"
-
-@fixture
-def order(first_entry):
-    return []
-
-@fixture({auto_use_kw}=True)
-def append_first(order, first_entry):
-    return order.append(first_entry)
-
-def test_string_only(order, first_entry):
-    assert order == [first_entry]
-
-def test_string_and_int(order, first_entry):
-    order.append(2)
-    assert order == [first_entry, 2]
-"#,
-    );
-
-    let case = IntegrationTestEnv::with_file("test_append.py", &test_code)?;
-
-    run_with_path_and_snapshot!(case,  @r"
-    success: true
-    exit_code: 0
-    ----- stdout -----
-    test result: ok. 2 passed; 0 failed; 0 skipped
-
-    ----- stderr -----
-    ");
-
-    Ok(())
-}
-
-#[rstest]
-fn test_fixture_order_respects_scope(
-    #[values("pytest", "karva")] framework: &str,
-) -> anyhow::Result<()> {
-    let auto_use_kw = get_auto_use_kw(framework);
-
-    let test_code = format!(
-        r"
-from {framework} import fixture
-
-data = {{}}
-
-@fixture(scope='module')
-def clean_data():
-    data.clear()
-
-@fixture({auto_use_kw}=True)
-def add_data():
-    data.update(value=True)
-
-def test_value(clean_data):
-    assert data.get('value')
-",
-    );
-
-    let case = IntegrationTestEnv::with_file("test_fixture_order_scope.py", &test_code)?;
-
-    run_with_path_and_snapshot!(case,  @r"
-    success: true
-    exit_code: 0
-    ----- stdout -----
-    test result: ok. 1 passed; 0 failed; 0 skipped
-
-    ----- stderr -----
-    ");
-
-    Ok(())
-}
-
 #[test]
 fn test_multiple_fixtures_not_found() -> anyhow::Result<()> {
     let case = IntegrationTestEnv::with_file(
@@ -2641,54 +2551,6 @@ fn test_multiple_fixtures_not_found() -> anyhow::Result<()> {
     error[fixture-not-found]: fixture 'c' not found
 
     test result: FAILED. 0 passed; 1 failed; 0 skipped
-
-    ----- stderr -----
-    ");
-
-    Ok(())
-}
-
-#[rstest]
-fn test_nested_generator_fixture(
-    #[values("pytest", "karva")] framework: &str,
-) -> anyhow::Result<()> {
-    let mut files = get_source_code("pass");
-    files.extend([
-        (
-            "tests/conftest.py".to_string(),
-            format!(
-                r"
-                from {framework} import fixture
-                from src import Calculator
-
-                @fixture
-                def calculator() -> Calculator:
-                    if 1:
-                        yield Calculator()
-                    else:
-                        yield Calculator()
-                ",
-            ),
-        ),
-        (
-            "tests/test_calculator.py".to_string(),
-            r"
-            from src import Calculator
-
-            def test_calculator(calculator: Calculator) -> None:
-                assert calculator.add(1, 2) == 3
-            "
-            .to_string(),
-        ),
-    ]);
-
-    let case = IntegrationTestEnv::with_files(files.iter().map(|(k, v)| (k.as_str(), v.as_str())))?;
-
-    run_with_path_and_snapshot!(case, &["-s"], @r"
-    success: true
-    exit_code: 0
-    ----- stdout -----
-    test result: ok. 1 passed; 0 failed; 0 skipped
 
     ----- stderr -----
     ");
